+++ test/inf-clojure-tests.el
@@ -173,7 +173,7 @@ is a string\")
   (it "provides all base features for clojure-family REPL types"
     (let ((base-features '(load doc source apropos ns-vars set-ns
                                 macroexpand macroexpand-1 arglists
-                                reload reload-all)))
+                                reload reload-all var-meta)))
       (dolist (repl-type '(clojure clojure-clr babashka node-babashka lein-clr))
         (dolist (feature base-features)
           (expect (inf-clojure--get-feature repl-type feature nil)
@@ -186,11 +186,13 @@ is a string\")
     (let ((clj-arglists (inf-clojure--get-feature 'clojure 'arglists nil))
           (bb-arglists (inf-clojure--get-feature 'babashka 'arglists nil)))
       (expect clj-arglists :to-equal bb-arglists)))
-  (it "provides reload and reload-all for all REPL types"
+  (it "provides reload, reload-all and var-meta for all REPL types"
     (dolist (repl-type '(clojure clojure-clr babashka node-babashka lein-clr cljs planck joker))
       (expect (inf-clojure--get-feature repl-type 'reload nil)
               :not :to-be nil)
       (expect (inf-clojure--get-feature repl-type 'reload-all nil)
+              :not :to-be nil)
+      (expect (inf-clojure--get-feature repl-type 'var-meta nil)
               :not :to-be nil)))
   (it "uses a different arglists catch clause for CLR REPL types"
     (let ((clj-arglists (inf-clojure--get-feature 'clojure 'arglists nil))